Move the status constants to own module
